Group triggers with the same name on overview screen according to their priority.
This allows to define several triggers with the same name but
different priorities. This patch performs such triggers sorting
according to their priority value. For example, we have three triggers
with name 'Lack of disk space on /' and the following keys and priorities:

1. INFORMATION: {host:vfs.fs.size[/,pfree].last(0)}<30
2. WARNING: {host:vfs.fs.size[/,pfree].last(0)}<20
3. HIGH: {host:vfs.fs.size[/,pfree].last(0)}<10
4. DISASTER: {host:vfs.fs.size[/,pfree].last(0)}<5

In overview screen we will see only one of these triggers. But when
free disk space percentage drops to 30%, we will see Information's
color brick. If this value drops to 20%, this brick changed color to
Warning and so on. Without this patch group was made, but color
changed randomly, not by priority.