[frontend] Allow screens to be defined on templates
Hi all

This is a preliminary patch, which allows screens to be defined
on templates. This allows one to define a screen, viewable for
multiple hosts, without having to define a screen per host, with
the same set of graphs in each screen.

There are still some issues, but it works for me.

Please give me any comments or suggestions.

Chris

http://www.zabbix.com/forum/archive/index.php/t-6669.html

Signed-off-by: Alexey Gladkov <legion@altlinux.org>